+++ src/storage/adapter/mod.rs
@@ -5,6 +5,8 @@
 #[cfg(feature = "rocksdb")]
 pub mod rocksdb;
 
+pub mod memory;
+
 use std::collections::HashMap;
 
 fn storage_err<E: ToString>(error: E) -> crate::Error {
@@ -26,33 +28,4 @@ pub trait StorageAdapter: std::fmt::Debug {
     async fn batch_set(&mut self, records: HashMap<String, String>) -> crate::Result<()>;
     /// Removes a record from the storage.
     async fn remove(&mut self, key: &str) -> crate::Result<()>;
-}
-
-#[derive(Debug, Default)]
-/// A storage adapter that stores data in memory.
-pub struct Memory(HashMap<String, String>);
-
-#[async_trait::async_trait]
-impl StorageAdapter for Memory {
-    async fn get(&self, key: &str) -> crate::Result<String> {
-        self.0.get(key).ok_or_else(|| storage_err(key)).cloned()
-    }
-
-    /// Saves or updates a record on the storage.
-    async fn set(&mut self, key: &str, record: String) -> crate::Result<()> {
-        self.0.insert(key.to_string(), record);
-        Ok(())
-    }
-
-    /// Batch write.
-    async fn batch_set(&mut self, records: HashMap<String, String>) -> crate::Result<()> {
-        self.0.extend(records.into_iter());
-        Ok(())
-    }
-
-    /// Removes a record from the storage.
-    async fn remove(&mut self, key: &str) -> crate::Result<()> {
-        self.0.remove(key);
-        Ok(())
-    }
 }
