+++ src/storage/constants.rs
@@ -4,8 +4,8 @@
 /// The default storage path.
 pub const DEFAULT_STORAGE_PATH: &str = "./storage";
 
-#[cfg(feature = "rocksdb")]
 /// The default RocksDB storage path.
+#[cfg(feature = "rocksdb")]
 pub(crate) const ROCKSDB_FOLDERNAME: &str = "walletdb";
 
 pub const fn default_storage_path() -> &'static str {
