+++ src/storage/manager.rs
@@ -17,11 +17,11 @@ use crate::{
 /// The storage used by the manager.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub(crate) enum ManagerStorage {
-    #[cfg(feature = "stronghold")]
     /// Stronghold storage.
+    #[cfg(feature = "stronghold")]
     Stronghold,
-    #[cfg(feature = "rocksdb")]
     /// RocksDB storage.
+    #[cfg(feature = "rocksdb")]
     Rocksdb,
     /// Storage backed by a Map in memory, for testing
     Memory,
@@ -73,8 +73,8 @@ pub(crate) async fn new_storage_manager(
     Ok(Arc::new(Mutex::new(storage_manager)))
 }
 
-#[derive(Debug)]
 /// Storage manager
+#[derive(Debug)]
 pub struct StorageManager {
     pub(crate) storage: Storage,
     // account indexes for accounts in the database
